working version of autoregressive correlation
