+++ requirements.txt
@@ -9,3 +9,5 @@ rpy2~=3.5.10
 r-base~=4.2.3
 bioconductor-limma~=3.54.0
 statsmodels~=0.13.5
+
+matplotlib~=3.7.1
